rename koiki.models to koiki.resources
