+++ .gitignore
@@ -1,4 +1,5 @@
 __pycache__/
+.env
 .envrc
 .coverage
 htmlcov/
